create rooms package, art package, relocate & refactor room and art classes
