+++ src/com/alienation/enginefiles/Engine.java
@@ -1,6 +1,7 @@
 package com.alienation.enginefiles;
 import com.alienation.coregamefiles.*;
 import com.alienation.coregamefiles.Character;
+import com.alienation.rooms.CapsuleRoom;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
