+++ src/com/alienation/gameart/Banner.java
@@ -1,6 +1,6 @@
-package com.alienation.coregamefiles;
+package com.alienation.gameart;
 
-import static com.alienation.coregamefiles.TextColors.*;
+import static com.alienation.gameart.TextColors.*;
 
 /**
  * Created by bradsmialek on Thu - 7/23/20 @ 1:20 PM
