+++ src/com/alienation/gameart/Win.java
@@ -1,7 +1,9 @@
-package com.alienation.coregamefiles;
+package com.alienation.gameart;
+
+import com.alienation.coregamefiles.Menu;
 
 import java.util.concurrent.TimeUnit;
-import static com.alienation.coregamefiles.TextColors.*;
+import static com.alienation.gameart.TextColors.*;
 
 /**
  * Created by bradsmialek on Wed - 7/29/20 @ 8:03 AM
